Creacion de objetos

+++ sistema-certificados-backend/src/main/java/com/agg/certificados/dtos/response/TypeDocumentResponseDto.java
@@ -1,7 +1,7 @@
 package com.agg.certificados.dtos.response;
 
 public class TypeDocumentResponseDto {
-    public int id_type_document;
+    public Long id_type_document;
     public String name;
     public String description;
     public boolean status;
