+++ sistema-certificados-backend/src/main/java/com/agg/certificados/entity/TypeDocument.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 public class TypeDocument {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public int id_type_document;
+    public Long id_type_document;
     public String name;
     public String description;
     public boolean status;
@@ -15,18 +15,18 @@ public class TypeDocument {
     public TypeDocument() {
     }
 
-    public TypeDocument(int id_type_document, String description, boolean status, String name) {
+    public TypeDocument(Long id_type_document, String description, boolean status, String name) {
         this.id_type_document = id_type_document;
         this.description = description;
         this.status = status;
         this.name = name;
     }
 
-    public int getId_type_document() {
+    public Long getId_type_document() {
         return id_type_document;
     }
 
-    public void setId_type_document(int id_type_document) {
+    public void setId_type_document(Long id_type_document) {
         this.id_type_document = id_type_document;
     }
